+++ src/main/resources/templates/frontend/artist-details.html
@@ -9,7 +9,7 @@
     <div id="noResultsMessageContainer"></div>
     <script th:inline="javascript">
       $(document).ready(function () {
-        artistDetails([[${discogsId}]])
+        artistDetails()
       })
     </script>
     </body>
