+++ src/main/resources/templates/frontend/fragments/layout.html
@@ -23,6 +23,7 @@
   <script th:src="@{/webjars/bootstrap/js/bootstrap.min.js}" src="#"></script>
   <script th:src="@{/webjars/lightbox//src/js/lightbox.js}" src="#"></script>
   <script th:src="@{/js/common.js}" src="#"></script>
+  <script th:src="@{/js/util.js}" src="#"></script>
   <script th:src="@{/js/search.js}" src="#"></script>
   <script th:src="@{/js/follow-artists.js}" src="#"></script>
   <script th:src="@{/js/artist-details.js}" src="#"></script>
