+++ src/main/resources/templates/frontend/search.html
@@ -9,7 +9,7 @@
     <div id="noResultsMessageContainer"></div>
     <script th:inline="javascript">
       $(document).ready(function () {
-        search([[${searchRequest}]])
+        search()
       })
     </script>
     </body>
